End of tidy

+++ client/src/app/messages/messages.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
-
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService:MessageService) { }
+  constructor(private messageService:MessageService ,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,13 +33,16 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+    this.confirmService.confirm('Confirm delete message','This cannot be undone').subscribe(result =>{
+          this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+      })
     })
   }
 
-  pageChange(event: any){
+  pageChanged(event: any){
     this.pageNumber = event.page;
     this.loadMessages();
   }
+  
 }
